Rough but functional first implementation
The files are deleted, but it would be nice if there was an escape hatch
in the form of a confirmation before deleting all the files. The layout
of the file metadata may be a bit much. And the size and times need to
be formatted. Lastly, it would be nice to not buffer the input so the
user doesn't have to hit the enter key to confirm.